Consolidate duplicated logic
