+++ libs/utils/decorate.js
@@ -1,4 +1,6 @@
-import { createTag } from './utils.js';
+import { createTag, createIntersectionObserver } from './utils.js';
+
+const ROOT_MARGIN = 1000;
 
 export function decorateButtons(el, size) {
   const buttons = el.querySelectorAll('em a, strong a, p > a strong');
@@ -229,7 +231,7 @@ export function handleObjectFit(bgRow) {
   });
 }
 
-export function getVideoIntersectionObserver() {
+function getVideoIntersectionObserver() {
   if (!window?.videoIntersectionObs) {
     window.videoIntersectionObs = new window.IntersectionObserver((entries) => {
       entries.forEach((entry) => {
@@ -250,7 +252,7 @@ export function getVideoIntersectionObserver() {
   return window.videoIntersectionObs;
 }
 
-export function applyInViewPortPlay(video) {
+function applyInViewPortPlay(video) {
   if (!video) return;
   if (video.hasAttribute('data-play-viewport')) {
     const observer = getVideoIntersectionObserver();
@@ -259,4 +261,26 @@ export function applyInViewPortPlay(video) {
     });
     observer.observe(video);
   }
+}
+
+export function turnAnchorIntoVideo({ hash, src, anchorTag }) {
+  const { dataset, parentElement } = anchorTag;
+  const attrs = getVideoAttrs(hash, dataset);
+  const video = `<video ${attrs}></video>`;
+  anchorTag.insertAdjacentHTML('afterend', video);
+  createIntersectionObserver({
+    el: parentElement,
+    options: { rootMargin: `${ROOT_MARGIN}px` },
+    callback: () => {
+      parentElement
+        .querySelector('video')
+        .appendChild(
+          createTag('source', { src, type: 'video/mp4' }),
+        );
+    },
+  });
+  const videoEl = parentElement.querySelector('video');
+  applyHoverPlay(videoEl);
+  applyInViewPortPlay(videoEl);
+  anchorTag.remove();
 }
